・APIクラスに通信のベースとなる関数を実装し、それを実行するようにしてsignupとloginを実装 ・Mainへ遷移後にloginができるかを試してみた
